add some test code and dependences
