+++ app/build.gradle
@@ -30,6 +30,11 @@ android {
             path "CMakeLists.txt"
         }
     }
+    testOptions {
+        // the android.jar file used to run unit tests does not contain any actual code
+        // (those APIs are provided only by the Android system image on a device)
+        unitTests.returnDefaultValues = true
+    }
 }
 
 dependencies {
@@ -41,8 +46,28 @@ dependencies {
     // add Anko lib, should use part when method full 4k
     compile "org.jetbrains.anko:anko:0.10.1"
 
+    // FOR TEST
+    androidTestCompile 'com.android.support.test:runner:0.5'
+    // Set this dependency to use JUnit 4 rules
+    androidTestCompile 'com.android.support.test:rules:0.5'
+    // Optional -- Hamcrest library
+    // androidTestCompile 'org.hamcrest:hamcrest-library:1.3'
+
+    /**
+     *  Optional -- UI testing with Espresso
+     *
+     *  If your build configuration includes a compile dependency for the support-annotations
+     *  library and an androidTestCompile dependency for the espresso-core library,
+     *  your build might fail due to a dependency conflict. To resolve, update your dependency
+     *  for espresso-core as follows:
+     */
     androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    testCompile 'junit:junit:4.12'
+    // Optional -- UI testing with UI Automator
+    // androidTestCompile 'com.android.support.test.uiautomator:uiautomator-v18:2.1.2'
+
+    // FOR Local Unit Tests
+    testCompile 'junit:junit:4.12' // Required -- JUnit 4 framework
+    testCompile 'org.mockito:mockito-core:2.2.15' // Optional -- Mockito framework
 }
