+++ app/src/androidTest/java/com/okaycamera/okcamera/MainActivityInstrumentationTest.kt
@@ -1,0 +1,72 @@
+/*
+ * Copyright (c) 2017, Jianguo Yang.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package com.okaycamera.okcamera
+
+import android.support.test.InstrumentationRegistry
+import android.support.test.espresso.Espresso.onView
+import android.support.test.espresso.action.ViewActions.click
+import android.support.test.espresso.assertion.ViewAssertions.matches
+import android.support.test.espresso.matcher.ViewMatchers.withId
+import android.support.test.espresso.matcher.ViewMatchers.withText
+import android.support.test.rule.ActivityTestRule
+import org.junit.Rule
+import android.support.test.runner.AndroidJUnit4
+import android.view.WindowManager
+import org.junit.Assert
+import org.junit.Before
+import org.junit.Test
+import org.junit.runner.RunWith
+
+
+@RunWith(AndroidJUnit4::class)
+class MainActivityInstrumentationTest {
+
+    @get:Rule
+    var mActivityRule: ActivityTestRule<*> = ActivityTestRule<Main>(Main::class.java)
+
+    /**
+     *  Use this annotation to specify a block of code that contains test setup operations.
+     *  The test class invokes this code block before each test.
+     *  You can have multiple @Before methods but the order in which the test class calls these
+     *  methods is not guaranteed.
+     */
+    @Before
+    fun setUp() {
+        val activity = mActivityRule.getActivity()
+        val wakeUpDevice = Runnable {
+            activity.getWindow().addFlags(WindowManager.LayoutParams.FLAG_TURN_SCREEN_ON or
+                    WindowManager.LayoutParams.FLAG_SHOW_WHEN_LOCKED or
+                    WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
+        }
+        activity.runOnUiThread(wakeUpDevice)
+    }
+
+    @Test
+    fun sayHello() {
+        onView(withText("Say hello!")).perform(click())
+
+        onView(withId(R.id.sample_text)).check(matches(withText("Hello, World!")))
+    }
+
+    @Test
+    @Throws(Exception::class)
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getTargetContext()
+        Assert.assertEquals("com.okaycamera.okcamera", appContext.packageName)
+    }
+}
