--- app/src/test/java/com/okaycamera/okcamera/ExampleUnitTest.kt
@@ -1,18 +1,0 @@
-package com.okaycamera.okcamera
-
-import org.junit.Test
-
-import org.junit.Assert.*
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-class ExampleUnitTest {
-    @Test
-    @Throws(Exception::class)
-    fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
-    }
-}
