+++ app/src/test/java/com/okaycamera/okcamera/FullTestSuite.java
@@ -1,0 +1,18 @@
+package com.okaycamera.okcamera;
+
+import android.test.suitebuilder.TestSuiteBuilder;
+
+import junit.framework.Test;
+import junit.framework.TestSuite;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see [Testing documentation](http://d.android.com/tools/testing)
+ */
+public class FullTestSuite extends TestSuite {
+
+    public FullTestSuite() {
+        super();
+    }
+}
