+++ build.gradle
@@ -21,7 +21,7 @@ buildscript {
 allprojects {
     repositories {
         jcenter()
-        maven { url 'https://maven.google.com' }
+        maven {url 'https://dl.google.com/dl/android/maven2/'}
         mavenCentral()
     }
 }
